requirements.txt
fastapi==0.*
uvicorn==0.*
python-multipart==0.*
websockets==15.*
python-dotenv==1.*
agno==1.*
pydantic==2.*
python-docx==1.*
PyPDF2==3.*
google-genai==1.*
ollama==0.*
openai==1.*